+++ Apps/Src/CommandLine.c
@@ -4,7 +4,7 @@
 #include <string.h>
 
 #include "BSP_UART.h"
-#include "CANbus.h"
+#include "CanBus.h"
 #include "Contactors.h"
 #include "Minions.h"
 #include "Pedals.h"
