+++ Apps/Src/DebugDump.c
@@ -1,6 +1,6 @@
 #include <errno.h>
 
-#include "CANbus.h"
+#include "CanBus.h"
 #include "Contactors.h"
 #include "Minions.h"
 #include "Pedals.h"
