+++ Apps/Src/ReadCarCan.c
@@ -5,9 +5,9 @@
  *
  */
 
-#include "ReadCarCAN.h"
+#include "ReadCarCan.h"
 
-#include "CANbus.h"
+#include "CanBus.h"
 #include "Contactors.h"
 #include "Display.h"
 #include "Minions.h"
